+++ camera/camera_driver.c
@@ -140,7 +140,7 @@ bool camera_driver_ctl(enum rarch_camera_ctl_state state, void *data)
 
             driver_ctl(RARCH_DRIVER_CTL_FIND_INDEX, &drv);
 
-            i = drv.len;
+            i         = (int)drv.len;
 
             if (i >= 0)
                camera_driver = (const camera_driver_t*)camera_driver_find_handle(i);
