+++ libretro-common/lists/string_list.c
@@ -269,7 +269,7 @@ int string_list_find_elem(const struct string_list *list, const char *elem)
    for (i = 0; i < list->size; i++)
    {
       if (strcasecmp(list->elems[i].data, elem) == 0)
-         return i+1;
+         return (int)(i + 1);
    }
 
    return false;
