+++ menu/cbs/menu_cbs_left.c
@@ -345,7 +345,7 @@ static int playlist_association_left(unsigned type, const char *label,
    if (next < 0)
    {
       if (wraparound)
-         next = list->count;
+         next = (int)(list->count);
       else
          next = 0;
    }
